remove PresetPatcherEnergy, make dedicated versions for prime and echoes
