+++ randovania/games/prime2/gui/preset_settings/__init__.py
@@ -11,6 +11,7 @@ if TYPE_CHECKING:
 def prime2_preset_tabs(editor: PresetEditor, window_manager: WindowManager) -> list[type[PresetTab]]:
     from randovania.games.prime2.gui.preset_settings.echoes_beam_configuration_tab import PresetEchoesBeamConfiguration
     from randovania.games.prime2.gui.preset_settings.echoes_dock_rando_tab import PresetEchoesDockRando
+    from randovania.games.prime2.gui.preset_settings.echoes_energy_tab import PresetEchoesEnergy
     from randovania.games.prime2.gui.preset_settings.echoes_goal_tab import PresetEchoesGoal
     from randovania.games.prime2.gui.preset_settings.echoes_item_pool_tab import EchoesPresetItemPool
     from randovania.games.prime2.gui.preset_settings.echoes_patches_tab import PresetEchoesPatches
@@ -20,7 +21,6 @@ def prime2_preset_tabs(editor: PresetEditor, window_manager: WindowManager) -> l
     from randovania.gui.preset_settings.generation_tab import PresetGeneration
     from randovania.gui.preset_settings.hints_tab import PresetHints
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
-    from randovania.gui.preset_settings.patcher_energy_tab import PresetPatcherEnergy
     from randovania.gui.preset_settings.trick_level_tab import PresetTrickLevel
 
     return [
@@ -30,7 +30,7 @@ def prime2_preset_tabs(editor: PresetEditor, window_manager: WindowManager) -> l
         PresetEchoesGoal,
         PresetLocationPool,
         EchoesPresetItemPool,
-        PresetPatcherEnergy,
+        PresetEchoesEnergy,
         PresetTeleportersPrime2,
         PresetEchoesStartingArea,
         PresetEchoesDockRando,
